*** empty log message ***

+++ include/intervaltypes.h
@@ -25,7 +25,7 @@ struct ltrecord
 };
 typedef set<MemoryTrace::iterator *, ltrecord> RecordList;
 
-typedef UINT8 TCreateList;
+typedef UINT16 TCreateList;
 static const TCreateList NOCREATE = 0x00;
 static const TCreateList CREATEEVENTS = 0x01;
 static const TCreateList CREATECOMMS = 0x02;
